[Fix] #30 - move files to UIComponents and modify enum code
